Reduce files count
